cloud_function/requirements.txt
google-cloud-bigquery
pandas
google-cloud-pubsub
google-cloud-storage